refactor: Update dependencies and styles for Geofisico page
This commit updates the package.json file to include the "plotly.js-dist" dependency at version 2.32.0. It also updates the geofisico.scss file to add a new background color variable, $background, with the value #070707. These changes are made to enhance the Geofisico page by adding interactive plots and improving the visual appearance with a darker background color.